transfer auth if we try to send a query to a non-working DC we're not logged into
We had a case where we have a non-logged-in connection to a DC (it
migrated us during sign in) and we later get a new query for that DC
before the connection timeout fires and the session is torn down. We
would get a 401 because we didn't have a valid authorization.