Fix a std::bad_weak_ptr exception
Inside tgl_connection_asio the close() method is called
by the destructor. As a result share_from_this() shouldn't
be called in it. In close() we only need a weak pointer to
remove it from the observer set inside tgl_state. The solution
is that keep a weak pointer captured early in the open() method
and use it after. Also remove mtproto_client::close() which
did nothing.