Prefer .mov over .qt for video/quicktime
This doesn't fix anything but just is a perference of extension for video/quicktime.
Also remove the LGPL license header for mime.types and mime-type.cpp. The mime.types
is in public domain without limit according to the original file. The mime-types.cpp
is actually a generated file with `xxd -i mime.types > mime-type.cpp`. There are
no LGPL license header for these original files.