Adjust query timer based on ack flag
Fixes: T1544

Sometimes the server could be slow for some types of queries
and it sends an ack but doesn't send a result immediately.
In this case we should restart the timer and give the server
more time to process the query.

Also adjust the timeout for low priority queries: send typing and mark read.