Fix random missing of video thumbnail
Fixes: T1442

The bug was inherit from the original upstream code. It checks the
random thumb_id for positive as a condition of sending thumbnail.
But apprently it can be negative. The fix is that checking it for
non-zero and making sure that it's not zero when we have thumbnail
to send.