Fix invokeAfterMsg
It didn't work becase previously we use the actual message id instead of
the mtprotocol message id which is the query id as the dependency.

This patch also improves the seqence number handling for secret chat. Now
we capture the in_seq_no and out_seq_no from a secret chat when a message
is created from the chat.