Implment receiving queue for organizing secret chat messages
Task: T1670

With this queue in place we can insure the receiving order of
secret chat messages base on the internal sequence number.
The next step (but not very urgently) is to persistent the
queue in case the app quit while there are still queued messages.

As a result the set_hole_detection_enabled() hack now has been removed.