message_id_update() callback can't tell chat id
When we get this update from the server we have definitely lost
the context about to which chat this message belongs. In the mean while
messsage_sent() callback can definitely tell the chat id since
we have the query which triggers it. As a result remove chat id
from message_id_update() and update documentations about it and
message_sent() callback.